Alias :root into section while adding theme
Replace :root selector into section by splitting out PostCSS plugin.
