+++ src/theme_set.js
@@ -9,6 +9,7 @@ import postcssPrintable, {
 } from './postcss/printable'
 import postcssPseudoPrepend from './postcss/pseudo_selector/prepend'
 import postcssPseudoReplace from './postcss/pseudo_selector/replace'
+import postcssRootFontSize from './postcss/root/font_size'
 import postcssRem from './postcss/root/rem'
 import postcssRootReplace from './postcss/root/replace'
 import Theme from './theme'
@@ -270,6 +271,7 @@ class ThemeSet {
         opts.inlineSVG && postcssAdvancedBackground,
         postcssPagination,
         postcssRootReplace,
+        postcssRootFontSize,
         postcssPseudoPrepend,
         postcssPseudoReplace(opts.containers, slideElements),
         opts.printable && postcssPrintablePostProcess,
